Added Palindrome checking, and a few modular funcs
Palindrome: checks to see if the provided number is a palindrome.
Modular Inverse: uses the Extended Euclidean Algorithm to find the
modular inverse of a number.  Assumes the number and the modulus are
coprime.
BinomMod: Uses Lucas' Theorem to compute the binomial coefficient modulo
a prime number.  Assumes provided modulus is prime.